+++ tests/cmd/test_build.lua
@@ -2,14 +2,15 @@ local levee = require("levee")
 local _ = levee._
 
 
-local script = [[
-local levee = require("levee")
-print("oh hai")
-]]
 
 
 return {
 	test_core = function()
+		local script = [[
+			local levee = require("levee")
+			io.stdout:write("oh hai")
+		]]
+
 		local tmp = _.path.Path:tmpdir()
 		defer(function() tmp:remove(true) end)
 
@@ -17,14 +18,13 @@ return {
 
 		local options = {
 			exe = tostring(tmp("foo")),
-			file = tostring(tmp("foo.lua")),
-		}
+			file = tostring(tmp("foo.lua")), }
 
 		require("levee.cmd").build.run(options)
 
 		local h = levee.Hub()
 		local child = h.process:spawn(options.exe)
-		assert.equal(child.stdout:reads(), "oh hai\n")
+		assert.equal(child.stdout:reads(), "oh hai")
 		child.done:recv()
 	end,
 }