+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * add support for sendto and recvfrom
 * move TCP connect to use a hub's poller, so connect is async
 * add support for TLS via libressl!
+* add a process:respawn convenience
 
 ### Deprecates